am26x : ecap : update examples on ecap emulation
Update the examples to use the ecap emulation free to be backward
compatible. Emulation Mode in Syscfg by default is not "emulation free"
and this will enable the emulation support to the peripheral. However,
the DebugP_log in the examples will effect the ecap counters as now the
emulation support will be present. the changes in the commit will avoid
that.

Fixes: MCUSDK-13341

Signed-off-by: Mihira Madhava Bollapragada <madhava@ti.com>